Initial commit
Griso provides a simple implementation for graphs with directed or
undirected edges. Edges may be named or unnamed. Based on this graph
implementation griso contains an algorithm for creating a canonical
labelling of a graph. If graphs are ambiguous griso creates allows the
user to iterate over all canonical labellings possible.

The canonical labellings are used to implement a graph isomorphism check.